cha - code refactor

+++ trunk/Serenna/sources/fr/hartland/serenna/core/connection/Connection.java
@@ -52,8 +52,8 @@ public class Connection
 	 * 
 	 * @return value computed during computation.
 	 */
-	public double getValue()
+	public Double getValue()
 	{
-		return inputNode.getValue();
+		return (Double) inputNode.getValue();
 	}
 }
